De-dup events in the base events queries (#74)
Previously we did this in the stg_ga4__events view, but that turned out to be incompatible with parition pruning, an important cost optimization.

Doing it in the "base" queries does not suffer from this problem because the incremental update process ensures that we are only querying one date partition at a time anyway, which is the right granularity to run de-duping at because the duplicate events have duplicate event_date_dt/event_timestamp values and therefore get delivered into the same partition.