Channel grouping fixes (#85)
* Reduce precedence of the referral channel grouping rule

eg. nz.search.yahoo.com is categorised as SOURCE_CATEGORY_SEARCH, but that got ignored as the medium is set to 'referral'.

This ordering matches the 'Channels for manual traffic' section of https://support.google.com/analytics/answer/9756891?hl=en.

* Add Paid Video channel grouping

* Assign dv360_display traffic in the Display channel

Need to prioritize this over the Paid Search rule because it sometimes has medium=cpc.